+++ Changelog.md
@@ -1,2 +1,29 @@
-# Change Log
-This file is maintained to store the versioning of the features we have put in each release.
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+### Added
+
+### Fixed
+
+### Changed
+
+### Removed
+
+## [0.0.1] - 2025-01-27
+
+### Added
+
+- This changelog file to hopefully serve as an evolving example of a standardized changelog.
+- Script for updating this changelog in the file `scripts/release/update_changelog.py`
+- Added base docs (README and contribution guidelines) for this repository
+- CODEOWNERS file to track mandatory reviewers of certain PRs
+
+-----
+[unreleased]: https://github.com/isaacchunn/wanderers/compare/v0.0.1...HEAD
+[0.0.1]: https://github.com/isaacchunn/wanderers/releases/tag/v0.0.1